renamed bidi to segementing
