+++ Lib/fontgoggles/misc/segmenting.py
